Error Handling https://book.leptos.dev/view/07_errors.html
In the last chapter, we saw that you can render Option<T>: in the None case, it will render nothing, and in the Some(T) case, it will render T (that is, if T implements IntoView).
You can actually do something very similar with a Result<T, E>. In the Err(_) case, it will render nothing. In the Ok(T) case, it will render the T.

<ErrorBoundary/>
An <ErrorBoundary/> is a little like the <Show/> component we saw in the last chapter.
If everything’s okay—which is to say, if everything is Ok(_)—it renders its children.
But if there’s an Err(_) rendered among those children, it will trigger the <ErrorBoundary/>’s fallback.

Let’s add an <ErrorBoundary/> to this example.